Support for older versions of Arista EOS
Some older versions of Arista EOS do not support some of the CLI
commands currently used by the Arista ML2 mechanism driver. Check
the version of EOS running on the switch when the driver starts
up, and use this to determine which commands are supported. Add an
option to log command errors at INFO level to avoid spurious error
messages.

Tested with Arista EOS 4.13.10M on a 7124SX switch.

Change-Id: I49d291cba44769e4c4ae3aee22e62b030e614f24
Closes-bug: 1474054