Fix port creation on shared networks
In create_port_precommit, the VM was only added to the Arista DB if the
port was associated to a network in the same tenant. But in the case of
a shared network, the port tenant and the network tenant can be
different, while the VM should still be added to the DB.

This commit ensures the VM is added to the DB when connected to a shared
network, using the same logic as in create_port_postcommit.

Change-Id: I295e6c87b460be4adb44e2275fa3008d3ce3eda7