Use requests library instead of jsonrpclib
This change modifies the Arista ML2 driver to remove the use of jsonrpclib and
use the requests library instead. An associated change now allows configuring
the timeout interval for the requests library.

Aditionally, we now attempt to check if the last known EOS master is still
available before executing an EOS command, allowing the driver to detect a
switchover and converge faster.

Change-Id: I5b58f23129dff8a299022dc5089ce7e43fc9a234
Closes-Bug:1524126