Add update_vm_host to db_lib
Re-adding update_vm_host to db_lib as it appears to have been
unintentionally removed. Also fixing test setup so that it runs
against neutron's kilo release instead of master.

This is not being backported because there is no new work or fix
in the master (or stable/liberty) for this. This method was accidently
dropped when stable/kilo was pulled of.

Change-Id: I6a8fd615f38d31b24cbf7e15add374e0e539c440
Closes-Bug: #1526952