Check that shared networks are VLAN networks before in Arista ML2
When we added support for shared networks, we introduced a bug that
can result in the arista ML2 driver reacting to a port-create for a
non-vlan network if the network is a shared network. With this
change the arista ML2 driver will only react to port creates for
shared networks if they are vlan networks.

Closes-Bug: #1553058

Conflicts:
	networking_arista/common/db_lib.py

Change-Id: I615431958e3a875270e4c90566d4ade56a715113
(Cherry picked from commit - 21e43c3d8bcf39a249cb692f53b1e1ed5aaa3d4d)