Added new option for filtering physical networks
This patch is the first in a series of patches to filter out processing
of port events based on a configuration option in ml2_arista_conf.ini.

In this patch, a new option (i.e. managed_physnet) is added to the
ml2_conf_arista.ini to filter out physical networks managed by
Arista’s mechanism driver.

If needed, the 'managed_physnets' option can be set to include a list of
physical networks to be managed by Arista’s mechanism driver.
The default setting does not exclude any physical network.

Change-Id: I4a9b183e64e93362d86ddb08f0d08e0b39f16e7b