Do not fail initialization when EOS not available
In the initialization of Arista ML2 mechanism driver, if registration with
EOS fails only log an error message and continue with the next step in the
initialization.
The registration will be performed in the sync method periodically.
A new flag is introduced which is an indication of availability of EOS to
Arista ML2 driver. In case of failure it is set to unavailable and when
registration succeeds it will be set to available.

Change-Id: I32218e2f2e3a97e66bf26a72a7da03cc4513d3ce
Closes-Bug: #1624501