Allow empty fields for Keystone.
Currently if there are empty identity auth fields for Keystone we send None and the API returns an error as these fields are required. Now we convert None to the empty string, "".

Closes-Bug: #1629405Allow empty fields for Keystone.

Currently if there are empty identity auth fields for Keystone we send None and the API returns an error as these fields are required. Now we convert None to the empty string, "".

Closes-Bug: #1629405Allow empty fields for Keystone.

Currently if there are empty identity auth fields for Keystone we send None and the API returns an error as these fields are required. Now we convert None to the empty string, "".

Closes-Bug: #1629405Allow empty fields for Keystone.

Currently if there are empty identity auth fields for Keystone we send None and the API returns an error as these fields are required. Now we convert None to the empty string, "".

Closes-Bug: #1629405Allow empty fields for Keystone.

Currently if there are empty identity auth fields for Keystone we send None and the API returns an error as these fields are required. Now we convert None to the empty string, "".

Closes-Bug: #1629405Allow empty fields for Keystone.

Currently if there are empty identity auth fields for Keystone we send None and the API returns an error as these fields are required. Now we convert None to the empty string, "".

Closes-Bug: #1629405Allow empty fields for Keystone.

Currently if there are empty identity auth fields for Keystone we send None and the API returns an error as these fields are required. Now we convert None to the empty string, "".

Closes-Bug: #1629405Allow empty fields for Keystone.

Currently if there are empty identity auth fields for Keystone we send None and the API returns an error as these fields are required. Now we convert None to the empty string, "".

Change-Id: I6d38d4c156c67385fe83059d2ab29ea9c60ef022
Closes-Bug: #1629405