Change passing session to context in segments db functions
We have a number of functions that expect to get session as one
the arguments. Passing session is not correct and this prevents
of usage new enginefacade which expect context to be passed and
session will be injected in current context.

This change is based on change in Neutron, so here all affected
places are fixed to pass context.

Depends-On: Ie1c54138ceaf9ac6f0674ad2786d4aaea9c80f73

Partially-Implements blueprint: enginefacade-switch

Change-Id: I878761a07d3ce826b8585877194851129a21c64e