Fixed a bug in segment creation and deletion
The segment db has 'segmentation_id' as a field where as the EOS API
expects segmentationId. While creating and deleting segments, the
segment list from the db was being passed to the API which resulted
in failures. This patch ensures that the segment data from the db
is massaged into the right format.

Change-Id: I35b5c7cf147ae2355dadfdeac2c8f2f69aa7107b