Add a manage_fabric configuration option
For fabric segments, the physnet is None. We should offer the
ability for a deployer to toggle whether arista driver acts as the
fabric driver while also offering the ability to have the arista
driver manage only a subset of VLAN segments. (I also cleaned up
some of our config descriptions.)

Change-Id: I4b2e6bd44e948dbeefefe0d344d551ee9e83e5dd