Fail VLAN allocations from outside the assigned range.
Recently, we stopped failing network_create_postcommit to allow
networks to be created even when CVX is unavailable. This broke
the arista vlan type driver functionality in the case of segment
creates with segmentation_id specified as the base vlan type driver
adds the segmentation id to the allocation range in this case.

This change checks whether the specified VLAN is in the allocation
range and fails the call if it isn't and calls the base vlan type
driver to finish the allocation if it is.

Change-Id: I723271c0a19e29f87bcf1e34e7413733478440ec