Add explicit check for rpc is not None
This change tests for the rpc value passed to the driver to be not None.
The explicity check does not change the logic and avoid a difference in
Python 2 vs Python 3 to resolve the boolean value of an object. Python 2
calls __nonzero__() to get boolean value of an object and Python 3
calls __boolean__().

Change-Id: I50a228840735bc1a7b0d857d0b28e01b4f836aef
(cherry picked from commit 6d2d1c34af61f3509666166d26892b3b4a26ddb3)