Add unordered list wrapper to help with validating mocked calls
This change simplifies mock.call() comparision logic by adding a wrapper to help
when comparing the contents of list of dictionaries. The wrapper treats the list
as an unordereed collection by disregarding the order of the elements when
checking for equality.  In production the lists are generated from database
queries resulting in unpredictable order.

Change-Id: I07a4003c2c4a54717654cecde76d28ad10854399