Use only 'binding:host_id' to detect port migration
The existing code fails to detect a port migration if the port status
is in 'DOWN' state. With this patch, detection of a port migration
relies only on change in 'binding:host_id'.
This patch also fixes failure of deleting a network due to orphan port
in the arista_provisioned_vms database by deleting all entries from the
table if they exist.

Change-Id: I4faa8c9aafc39dce9e0773fcf2ff4836b86080dd