Fix shared network port synchronization support
Shared network ports are saved under the network owner
tenant in the arista db. However, when we go to the
neutron db to find the device owner, we filter by tenant id.
Since the tenant id in the neutron db does not match the tenant
id in the neutron db, we do not find the associated neutron port
and therefore skip over any attempt to provision the port on CVX.
The fix is to simply query all neutron ports rather than filtering
by tenant.

Change-Id: Ib73b7ce8b3e3501e7c0c730ec11ff72e970d537d