Fix DHCP port deletes during network deletes
When a network is deleted, a possible race can occur where the
network_delete_precommit is processed before the port_delete_postcommit
for the DHCP port. Our delete port logic only sends the delete
port rpc call if the network exists in our db, which it doesn't if the
network_delete_precommit wins the race.

The fix is to just unconditionally send the RPC delete regardless of the
network's presence in our db.

Change-Id: Ibc531b1a18e4ddc362cb67192246d2f4a9a16226