Pass resources from Arista Mech Driver to Sync Worker for provisioning
This change actually utilizes the provisioning queue to provision
resources from the Arista mechanism driver. With this change, all
provisioning is now done by the sync worker.

A large part of this change is a rework of test_mechanism_arista.py
to validate the new provisioning mechanism. The new tests are closer to
end-to-end tests than unit tests. The tests now initialize neutron's ML2
plugin, load the Arista and OVS mech drivers and starts their workers.
Networks and ports are then created via calls to the neutron ML2 plugin,
which will then call the arista driver with context formatted correctly
(rather than the fake context we previously used.) The arista driver
will populate the provision queue in response to these calls and the
sync worker will eventually provision a MockCvx instance. The use of a
MockCvx instance is the only thing that really separates this from a
full end-to-end test as it doesn't actually validate the rpc calls it
sees in the way that the JSON api on CVX would.

I've also added some code to the test to enable profiling (it's disabled
by default) to simplify decisions about any future optimizations that we
may elect to pursue.

Change-Id: Ia544773dc1b88ed6aecce591841ca473f29f02b9