Unconditionally setting the authentication header
It is possible for the password to be empty. In such cases, an EAPI
request would fail with error 401 - No authentication header found. This
change adds the authentication header to the requests' session even if
username or password is empty. If the credentials are not correct, then
the request fails with:
401 - Unable to authenticate user: Bad username/password combination
which is now logged.

Change-Id: Ica45ce79dda546825565450f4d3f71103d8dd082
(cherry picked from commit 151b06d5e2a97284ac8ac577a470a5b1e23cbe04)