Prevent sync thread from crashing
Both the eapi and json rpc can reraise exceptions if they are not
one of the many exceptions that we expect and handle more gracefully.
Reraising these exceptions can crash the sync thread. In ocata and
earlier releases, there is no mechanism to revive the thread, so
sync remains broken until neutron is restarted.

The proposed fix is to indiscriminately catch and log all exceptions
in the sync thread.

Change-Id: I32d5da7cc8c87c0fee248eae31ae966c9cb0e424