Improve Security Group support
This implements the use of dynamic ACLs to ensure that stale
security groups and rules do not persist on TORs.

It also adds logic to query TORs for dynamic ACLs and ACL bindings.
This allows us to cleanup any stale config that might be present as a
result of intermittent connection loss. It also allows to only send
config that may be missing, rather than reconfiguring everything on
each sync.

Change-Id: I2beaa0b182cf054e2846145322b083781467d018