Align baremetal EGRESS/INGRESS meaning with VM meaning
When a SG is applied to a VM, ingress refers to traffic flowing
into a VM and egress refers to traffic flowing out.

In the baremetal case, traffic flowing out of a switchport is
flowing into the the bare metal. Therefore, INGRESS SG rules
should be applied as 'out' ACLs and EGRESS rules as 'in' ACLs.

The names of the rules have been adjusted to match the OpenStack
meaning on the rules for clarity.

Change-Id: I1e8937cf6a4290782e0377f9061b72040d5b5716