Increase default command timeout
In larger deployments 'show openstack config ...' can
take longer than 10 seconds to run, resulting in a
connection timeout. (It's unclear whether this affects
JSON GETs/POSTs as well.) This increases the default to 60
seconds as a longer default seems to have little downside
and will avoid timeouts in larger deployments.

Change-Id: I1c6caf5339ce57af33cb552065d53c922863f899