Fix to prevent continuous sync failures after a failed sync
When beginning a sync, we set current_sync_name in the JSON rpc after
a successful POST to the sync endpoint. The only place we reset to the
current_sync_name is in the sync_end(). Therefore, any failure in sync
will result in current_sync_name not being cleared. A future attempt to
begin sync will fail as a result since a POST to sync with a sync_id
specified returns an error if a sync is not in progress with that sync ID.

The fix is to clear the current_sync_name when beginning a new sync attempt.

We also need to update CVX's sync interval if it's not set to avoid immediate
sync timeout.

Another issue is that we're prematurely setting the local uuid to cvx uuid,
preventing full syncs if the full sync fails.

This also improves logging for arista_resources.

Change-Id: I5adc6f0f420f1d33e05a936d8ca4227c44414d4c