A number of fixes for the driver rewrite
1. When beginning a sync, we set current_sync_name in the JSON rpc after
a successful POST to the sync endpoint. The only place we reset to the
current_sync_name is in the sync_end(). Therefore, any failure in sync
will result in current_sync_name not being cleared. A future attempt to
begin sync will fail as a result since a POST to sync with a sync_id
specified returns an error if a sync is not in progress with that sync ID.
The fix is to clear the current_sync_name when beginning a new sync attempt.

2. Update CVX's sync interval if it's not set to avoid immediate sync timeout.

3. Don't prematurely set the local uuid to cvx uuid, preventing full syncs
if the full sync fails.

4. We were querying CVX before beginning sync, we now do that after sync
begins.

5. Get rid of threading.Event for sync_worker communication. Neutron starts
the sync worker in another process, so it doesn't work. multiprocessing.Queue
uses sockets and fds to implement an interprocess Queue that does wakeup the
sync worker within the other process. It however does not wakeup a thread
within the same process, so use eventlet.queue.LightQueue for testing.

6. Don't call delete_port_resources for a transition from DOWN -> DOWN

7. Fix network name and instance hostId to match the JSON API model

8. Improve logging for arista_resources and mechanism_arista.

9. Don't sync tenants with id ''

Change-Id: I5adc6f0f420f1d33e05a936d8ca4227c44414d4c