Delete resources in reverse sync order
Dependent resources should be deleted before the resources they
depend on to allow CVX to better handle the DELETEs. This avoids
situations such as where we delete a network before its segments
causing the segment DELETEs to fail as a result of being unable
to find the network in the region's network collection.

Change-Id: I1f1fa2d199b45bed52c549731580c06a306a3231