Fix JSON API recovery from sync failure
When beginning a sync, we set current_sync_name in the JSON rpc after
a successful POST to the sync endpoint. The only place we reset to the
current_sync_name is in the sync_end(). Therefore, any failure in sync
will result in current_sync_name not being cleared. A future attempt to
begin sync will fail as a result since a POST to sync with a sync_id
specified returns an error if a sync is not in progress with that sync
ID. The fix is to clear the current_sync_name when beginning a new sync
attempt.

Change-Id: I2bcaa4e2a1467b1913cd95aedcd169c5e3dff379