Allow arista_vlan type driver to startup when CVX is not reachable
When the arista_vlan type driver is enabled, but CVX is not reachable,
Neutron cannot be started as the parsing of available VLANs hits an
exception. This change causes available VLANs to be an empty list
at startup, which will prevent VLAN network creation until CVX is reachable,
but will not interfere with Neutron startup.

Change-Id: I100a62da7782023e9acb167662c2a5159bedf338